# string.c: Optimize String#concat when argc is 1