--- a/string.c
+++ b/string.c
@@ -2935,7 +2935,9 @@ rb_str_concat_multi(int argc, VALUE *argv, VALUE str)
 {
     str_modifiable(str);
 
-    if (argc > 0) {
+    if (argc == 1) {
+	return rb_str_concat(str, argv[0]);
+    } else if (argc > 1) {
 	int i;
 	VALUE arg_str = rb_str_tmp_new(0);
 	rb_enc_copy(arg_str, str);